requirements.txt
tornado
boto
python-magic
filechunkio
